requirements.txt
streamlit==1.28.1
pandas==2.1.2
plotly==5.18.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
openpyxl==3.1.2
psycopg2-binary==2.9.9
python-decouple==3.8
PyYAML==6.0.1
requests==2.31.0
polars==0.19.12
pandera==0.17.2
great-expectations==0.17.19
multimethod==1.9.1
typeguard==4.1.5
typing-extensions==4.8.0